refactor component imports to use relative paths and update component aliases
