+++ app/cart/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Cart from "../../components/Cart/Cart";
+import Cart from "../components/Cart/Cart";
 
 export default function CartPage() {
   return (
