+++ app/layout.tsx
@@ -1,6 +1,7 @@
+import { Metadata } from "next";
 import { Arimo } from "next/font/google";
-import Footer from "../components/layouts/Footer";
-import Header from "../components/layouts/Header";
+import Footer from "./components/layouts/Footer";
+import Header from "./components/layouts/Header";
 import "./globals.css";
 
 const arimo = Arimo({
